+++ client/src/mocks/data/MyPage/commentListMock.ts
@@ -1,118 +1,71 @@
-import { MyCommentType } from '@type/space.type';
+import {MyCommentType} from '@type/space.type';
 
 const commentListMock: MyCommentType[] = [
-  {
-    id: 1,
-    content:
-      '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
-    createDate: String(new Date()),
-    spaceTitle: 'SPACE 401',
-    writer: {
-      nick: '최고양',
-      url: 'https://s3-alpha-sig.figma.com/img/02f9/c229/6e3726a13538fa71aab55c92c8eb950a?Expires=1696809600&Signature=F3aa5ljj2H4gtM9d~joIHMN1CZGJC8Nu5bomeA-a98TgrpJ20rsyPOOstLhG57YbuPX0GkuY5togwKHcSWUkYqi1w~aqrV-dBP4Sqhr2pC7JAUeSbwLLoGCQuzbIqEsKq4stgV3x24Xbk81asoGubOrYkzAB2CFWd5MoMMCbfyG3nSoy~dzhKQbQbAjUWmtaly-xlcdZe6xVbGv5GXZBh8rXNVTFvg7dzNRRW3GQIkZz9~nK72s9E6GAi8D5dhESRslGqOAK5xsC04EmFc4Mlk09WafNEVglXWSidDBfX5c4OU-4JEjqU9ZwFIV1h-DwQZ-mJ93eJfXhuTWaIA32pg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
+    {
+        postId: 1,
+        commentId: 1,
+        spaceTitle: 'space401',
+        postTitle: '당신은 전문가',
+        postContent: '블라디미르 푸틴 러시아 대통령이 김정은 북한 국무위원장에게 “당신은 전문가”라고 말하는 영상이 공개됐다.\n' +
+            '러시아 국영방송 ‘로시야’의 파벨 자루빈 기자는 지난 15일(현지시각) 텔레그램에 ‘아직 아무도 본 적 없는 대화’라며 푸틴 대통령과 김 위원장의 미공개 영상을 올렸다. ...',
+        commentContent: '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
+        commentCreateDate: String(new Date()),
+        postCreateDate: String(new Date()),
+        mainImgUrl:
+            'https://s3-alpha-sig.figma.com/img/2297/aa17/07a3bbdc34704149870515161718468f?Expires=1696809600&Signature=WrsvWInolRDXhnK9hYHk0ldk8VgrKN80DibBnI5~H-ZO3d3jksTXHLWHOmv4EpmWz2mqfS0vnRK-xHUsjkpJQrw29p6DBYeWz5ZeMZ0fkVaisAbDNLgLHJPqyzPYaHnG9wLd6HEbSBQ9T0aSFg1tCBuSxnjzj7H2KqN1cfqvZhgvzT-gZ-iCLZk0dTiRpQ0v4TaFq4JsuL2~Yy5y5tF2JvMFes7ze~ietqnV2~kSeLTpUyg2aw7f9olI6XYmhGX394PFADGh~6m4tlNjEkzl3NcYqivqgAUKb46mpC1xakdDLsCmXY63aoKeesL03Raeij7YGIzAJMeCbBQOTNaZGQ__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
     },
-    mainImgUrl:
-      'https://s3-alpha-sig.figma.com/img/2297/aa17/07a3bbdc34704149870515161718468f?Expires=1696809600&Signature=WrsvWInolRDXhnK9hYHk0ldk8VgrKN80DibBnI5~H-ZO3d3jksTXHLWHOmv4EpmWz2mqfS0vnRK-xHUsjkpJQrw29p6DBYeWz5ZeMZ0fkVaisAbDNLgLHJPqyzPYaHnG9wLd6HEbSBQ9T0aSFg1tCBuSxnjzj7H2KqN1cfqvZhgvzT-gZ-iCLZk0dTiRpQ0v4TaFq4JsuL2~Yy5y5tF2JvMFes7ze~ietqnV2~kSeLTpUyg2aw7f9olI6XYmhGX394PFADGh~6m4tlNjEkzl3NcYqivqgAUKb46mpC1xakdDLsCmXY63aoKeesL03Raeij7YGIzAJMeCbBQOTNaZGQ__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    postId: 10,
-  },
-  {
-    id: 2,
-    content:
-      '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
-    createDate: String(new Date()),
-    spaceTitle: 'SPACE 401',
-    writer: {
-      nick: '최고양',
-      url: 'https://s3-alpha-sig.figma.com/img/02f9/c229/6e3726a13538fa71aab55c92c8eb950a?Expires=1696809600&Signature=F3aa5ljj2H4gtM9d~joIHMN1CZGJC8Nu5bomeA-a98TgrpJ20rsyPOOstLhG57YbuPX0GkuY5togwKHcSWUkYqi1w~aqrV-dBP4Sqhr2pC7JAUeSbwLLoGCQuzbIqEsKq4stgV3x24Xbk81asoGubOrYkzAB2CFWd5MoMMCbfyG3nSoy~dzhKQbQbAjUWmtaly-xlcdZe6xVbGv5GXZBh8rXNVTFvg7dzNRRW3GQIkZz9~nK72s9E6GAi8D5dhESRslGqOAK5xsC04EmFc4Mlk09WafNEVglXWSidDBfX5c4OU-4JEjqU9ZwFIV1h-DwQZ-mJ93eJfXhuTWaIA32pg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
+    {
+        postId: 2,
+        commentId: 2,
+        spaceTitle: 'space401',
+        postTitle: '당신은 전문가',
+        postContent: '블라디미르 푸틴 러시아 대통령이 김정은 북한 국무위원장에게 “당신은 전문가”라고 말하는 영상이 공개됐다.\n' +
+            '러시아 국영방송 ‘로시야’의 파벨 자루빈 기자는 지난 15일(현지시각) 텔레그램에 ‘아직 아무도 본 적 없는 대화’라며 푸틴 대통령과 김 위원장의 미공개 영상을 올렸다. ...',
+        commentContent: '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
+        commentCreateDate: String(new Date()),
+        postCreateDate: String(new Date()),
+        mainImgUrl:
+            'https://s3-alpha-sig.figma.com/img/2297/aa17/07a3bbdc34704149870515161718468f?Expires=1696809600&Signature=WrsvWInolRDXhnK9hYHk0ldk8VgrKN80DibBnI5~H-ZO3d3jksTXHLWHOmv4EpmWz2mqfS0vnRK-xHUsjkpJQrw29p6DBYeWz5ZeMZ0fkVaisAbDNLgLHJPqyzPYaHnG9wLd6HEbSBQ9T0aSFg1tCBuSxnjzj7H2KqN1cfqvZhgvzT-gZ-iCLZk0dTiRpQ0v4TaFq4JsuL2~Yy5y5tF2JvMFes7ze~ietqnV2~kSeLTpUyg2aw7f9olI6XYmhGX394PFADGh~6m4tlNjEkzl3NcYqivqgAUKb46mpC1xakdDLsCmXY63aoKeesL03Raeij7YGIzAJMeCbBQOTNaZGQ__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
     },
-    mainImgUrl:
-      'https://s3-alpha-sig.figma.com/img/f588/8cb0/d0c9a3f8fccf505547ead7739eb048ae?Expires=1696809600&Signature=c24~Zs3dhbDvqZ25dl7IBN91UEmqca6aUjJfEEJDnukwg-Q2W289ZJi-AmRloCx63FntiD3ieL69qPBeVijMMZ57wXhRr6nbenkouHtzEX3rBp4kT4cuLjjGTgI3rwKs7lxXST-CIisD8VodwVoftIz0jlQROkNsPq4WtFio8VY8Q2il21CGAaTS2WZVkLlnlsEM23BVbR5suEi47ptPlWfYIpRg~GfCcM5Nhqqtgnjnr8465SSVrfwouJF3j6MXdFOOJ5wT39Va~SEjGannm-GSLugr5yh9XLPQIoLUk0Qu53wVzDMbgmHUMN1JocV6-MYjsM6xiZlY2u9PCbglBg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    postId: 10,
-  },
-  {
-    id: 3,
-    content:
-      '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
-    createDate: String(new Date()),
-    spaceTitle: 'SPACE 401',
-    writer: {
-      nick: '최고양',
-      url: 'https://s3-alpha-sig.figma.com/img/02f9/c229/6e3726a13538fa71aab55c92c8eb950a?Expires=1696809600&Signature=F3aa5ljj2H4gtM9d~joIHMN1CZGJC8Nu5bomeA-a98TgrpJ20rsyPOOstLhG57YbuPX0GkuY5togwKHcSWUkYqi1w~aqrV-dBP4Sqhr2pC7JAUeSbwLLoGCQuzbIqEsKq4stgV3x24Xbk81asoGubOrYkzAB2CFWd5MoMMCbfyG3nSoy~dzhKQbQbAjUWmtaly-xlcdZe6xVbGv5GXZBh8rXNVTFvg7dzNRRW3GQIkZz9~nK72s9E6GAi8D5dhESRslGqOAK5xsC04EmFc4Mlk09WafNEVglXWSidDBfX5c4OU-4JEjqU9ZwFIV1h-DwQZ-mJ93eJfXhuTWaIA32pg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
+    {
+        postId: 3,
+        commentId: 3,
+        spaceTitle: 'space401',
+        postTitle: '당신은 전문가',
+        postContent: '블라디미르 푸틴 러시아 대통령이 김정은 북한 국무위원장에게 “당신은 전문가”라고 말하는 영상이 공개됐다.\n' +
+            '러시아 국영방송 ‘로시야’의 파벨 자루빈 기자는 지난 15일(현지시각) 텔레그램에 ‘아직 아무도 본 적 없는 대화’라며 푸틴 대통령과 김 위원장의 미공개 영상을 올렸다. ...',
+        commentContent: '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
+        commentCreateDate: String(new Date()),
+        postCreateDate: String(new Date()),
+        mainImgUrl:
+            'https://s3-alpha-sig.figma.com/img/2297/aa17/07a3bbdc34704149870515161718468f?Expires=1696809600&Signature=WrsvWInolRDXhnK9hYHk0ldk8VgrKN80DibBnI5~H-ZO3d3jksTXHLWHOmv4EpmWz2mqfS0vnRK-xHUsjkpJQrw29p6DBYeWz5ZeMZ0fkVaisAbDNLgLHJPqyzPYaHnG9wLd6HEbSBQ9T0aSFg1tCBuSxnjzj7H2KqN1cfqvZhgvzT-gZ-iCLZk0dTiRpQ0v4TaFq4JsuL2~Yy5y5tF2JvMFes7ze~ietqnV2~kSeLTpUyg2aw7f9olI6XYmhGX394PFADGh~6m4tlNjEkzl3NcYqivqgAUKb46mpC1xakdDLsCmXY63aoKeesL03Raeij7YGIzAJMeCbBQOTNaZGQ__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
     },
-    mainImgUrl:
-      'https://s3-alpha-sig.figma.com/img/8dd6/f06c/415ef5e1ab1c32d5c01945d4d4e0d65c?Expires=1696809600&Signature=ezcF1Xwhp3bEtqa-XrklYpaaebgpucql8uu831T0MeiYIZgyCh3Gxb~ZM-hCH2pNmhcLKF9BWcQ7lyg0LfnWZ5-ANfD90D1gSAef~TugBGlyZKIwl4Z4bpckLw13DmMTTlWhHVuKL2fBei0xvqq0~KmgZXmN8yKH7QX7Y0RiQNoW8lK7RjiOru~2lMR0Kli6lUj-KkDY2HFS7UiFwCbMp2IeAJpJ7S9XKA76xPJq7kRJ9VZjBVDTR4mYShA3vKnLb8b38aFvYiGAL9~ddVQhAABTemu9XkfHy1c1GGexqtHujrswRv0Ejqv7GXtcXYre9fBNk~aXcCDc2Edd09IsJQ__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    postId: 10,
-  },
-  {
-    id: 4,
-    content:
-      '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
-    createDate: String(new Date()),
-    spaceTitle: 'SPACE 401',
-    writer: {
-      nick: '최고양',
-      url: 'https://s3-alpha-sig.figma.com/img/02f9/c229/6e3726a13538fa71aab55c92c8eb950a?Expires=1696809600&Signature=F3aa5ljj2H4gtM9d~joIHMN1CZGJC8Nu5bomeA-a98TgrpJ20rsyPOOstLhG57YbuPX0GkuY5togwKHcSWUkYqi1w~aqrV-dBP4Sqhr2pC7JAUeSbwLLoGCQuzbIqEsKq4stgV3x24Xbk81asoGubOrYkzAB2CFWd5MoMMCbfyG3nSoy~dzhKQbQbAjUWmtaly-xlcdZe6xVbGv5GXZBh8rXNVTFvg7dzNRRW3GQIkZz9~nK72s9E6GAi8D5dhESRslGqOAK5xsC04EmFc4Mlk09WafNEVglXWSidDBfX5c4OU-4JEjqU9ZwFIV1h-DwQZ-mJ93eJfXhuTWaIA32pg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
+    {
+        postId: 4,
+        commentId: 4,
+        spaceTitle: 'space401',
+        postTitle: '당신은 전문가',
+        postContent: '블라디미르 푸틴 러시아 대통령이 김정은 북한 국무위원장에게 “당신은 전문가”라고 말하는 영상이 공개됐다.\n' +
+            '러시아 국영방송 ‘로시야’의 파벨 자루빈 기자는 지난 15일(현지시각) 텔레그램에 ‘아직 아무도 본 적 없는 대화’라며 푸틴 대통령과 김 위원장의 미공개 영상을 올렸다. ...',
+        commentContent: '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
+        commentCreateDate: String(new Date()),
+        postCreateDate: String(new Date()),
+        mainImgUrl:
+            'https://s3-alpha-sig.figma.com/img/2297/aa17/07a3bbdc34704149870515161718468f?Expires=1696809600&Signature=WrsvWInolRDXhnK9hYHk0ldk8VgrKN80DibBnI5~H-ZO3d3jksTXHLWHOmv4EpmWz2mqfS0vnRK-xHUsjkpJQrw29p6DBYeWz5ZeMZ0fkVaisAbDNLgLHJPqyzPYaHnG9wLd6HEbSBQ9T0aSFg1tCBuSxnjzj7H2KqN1cfqvZhgvzT-gZ-iCLZk0dTiRpQ0v4TaFq4JsuL2~Yy5y5tF2JvMFes7ze~ietqnV2~kSeLTpUyg2aw7f9olI6XYmhGX394PFADGh~6m4tlNjEkzl3NcYqivqgAUKb46mpC1xakdDLsCmXY63aoKeesL03Raeij7YGIzAJMeCbBQOTNaZGQ__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
     },
-    mainImgUrl:
-      'https://s3-alpha-sig.figma.com/img/2b8d/50ac/ac8a45bd0867ec2c987f66839e894f69?Expires=1696809600&Signature=FzZFyzu8vDdIXvyM3bsGOgi5IVbvcNGEyW4dukoDBNFlDrP5WD~8riu~J5IzTAc4Wmh0O3EzjvQHu3TL1Y1cByDWL6N7xwce7uVJR2BQY5Tog2stZ8zF4FQizDGkkKWBgtR07Qdq3XccwmvXvIvpUZAupKD~B~U7v0~~GIZ-eZYKMOg29AikoAjIuxMC6bsCyiQFx9ii5SLaKnBu057E8i3h-1L6F3q52b4OTzejk53mcf1CXRQNSWw7J05W~9RzcvaOAZ0JUxVEmyES3U72J5l~b-RIq5zFpKoCcZRUxre01DWjsBSwwQBNxXWYgNSVMcrJ55od8V1NErQRX3~sEQ__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    postId: 10,
-  },
-  {
-    id: 5,
-    content:
-      '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
-    createDate: String(new Date()),
-    spaceTitle: 'SPACE 401',
-    writer: {
-      nick: '최고양',
-      url: 'https://s3-alpha-sig.figma.com/img/02f9/c229/6e3726a13538fa71aab55c92c8eb950a?Expires=1696809600&Signature=F3aa5ljj2H4gtM9d~joIHMN1CZGJC8Nu5bomeA-a98TgrpJ20rsyPOOstLhG57YbuPX0GkuY5togwKHcSWUkYqi1w~aqrV-dBP4Sqhr2pC7JAUeSbwLLoGCQuzbIqEsKq4stgV3x24Xbk81asoGubOrYkzAB2CFWd5MoMMCbfyG3nSoy~dzhKQbQbAjUWmtaly-xlcdZe6xVbGv5GXZBh8rXNVTFvg7dzNRRW3GQIkZz9~nK72s9E6GAi8D5dhESRslGqOAK5xsC04EmFc4Mlk09WafNEVglXWSidDBfX5c4OU-4JEjqU9ZwFIV1h-DwQZ-mJ93eJfXhuTWaIA32pg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
+    {
+        postId: 5,
+        commentId: 5,
+        spaceTitle: 'space401',
+        postTitle: '당신은 전문가',
+        postContent: '블라디미르 푸틴 러시아 대통령이 김정은 북한 국무위원장에게 “당신은 전문가”라고 말하는 영상이 공개됐다.\n' +
+            '러시아 국영방송 ‘로시야’의 파벨 자루빈 기자는 지난 15일(현지시각) 텔레그램에 ‘아직 아무도 본 적 없는 대화’라며 푸틴 대통령과 김 위원장의 미공개 영상을 올렸다. ...',
+        commentContent: '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
+        commentCreateDate: String(new Date()),
+        postCreateDate: String(new Date()),
+        mainImgUrl:
+            'https://s3-alpha-sig.figma.com/img/2297/aa17/07a3bbdc34704149870515161718468f?Expires=1696809600&Signature=WrsvWInolRDXhnK9hYHk0ldk8VgrKN80DibBnI5~H-ZO3d3jksTXHLWHOmv4EpmWz2mqfS0vnRK-xHUsjkpJQrw29p6DBYeWz5ZeMZ0fkVaisAbDNLgLHJPqyzPYaHnG9wLd6HEbSBQ9T0aSFg1tCBuSxnjzj7H2KqN1cfqvZhgvzT-gZ-iCLZk0dTiRpQ0v4TaFq4JsuL2~Yy5y5tF2JvMFes7ze~ietqnV2~kSeLTpUyg2aw7f9olI6XYmhGX394PFADGh~6m4tlNjEkzl3NcYqivqgAUKb46mpC1xakdDLsCmXY63aoKeesL03Raeij7YGIzAJMeCbBQOTNaZGQ__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
     },
-    mainImgUrl:
-      'https://s3-alpha-sig.figma.com/img/7be8/ff29/019d70c0ff45a640c3831bbc42c39a14?Expires=1696809600&Signature=itD~XLBXw5SNqpuSUKfhdBSwz0fD6sLnNvfVyHy2yXgIRlP7CeaH2ogR7JZ0fsW-FOVqolh1FRCCGT3hFinOabhyiRJZzosO22QQQ6v0Swd2PO0FfBUGDIGlYAIHAnZNWI17O4z-GRcR9dmn16DcVjHVPlJ5D3ecstSUUcIWKA84RBJyjg5GUATrk5dMibPzrkHRqi2CgJ-TlNAC-qjZTYbU6ssRmSVPL6T9AfhCYWB3sHXZjHWt4Utzsk5MhW-F2yzj0NO9ew~sUO52pST0hwPfd3yFTjyI0ovqNhwkQ4L~G5CyIi0hmggwZ~gnScFuKldzb-DU7l2~0gRsS35YKw__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    postId: 10,
-  },
-  {
-    id: 6,
-    content:
-      '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
-    createDate: String(new Date()),
-    spaceTitle: 'SPACE 401',
-    writer: {
-      nick: '최고양',
-      url: 'https://s3-alpha-sig.figma.com/img/02f9/c229/6e3726a13538fa71aab55c92c8eb950a?Expires=1696809600&Signature=F3aa5ljj2H4gtM9d~joIHMN1CZGJC8Nu5bomeA-a98TgrpJ20rsyPOOstLhG57YbuPX0GkuY5togwKHcSWUkYqi1w~aqrV-dBP4Sqhr2pC7JAUeSbwLLoGCQuzbIqEsKq4stgV3x24Xbk81asoGubOrYkzAB2CFWd5MoMMCbfyG3nSoy~dzhKQbQbAjUWmtaly-xlcdZe6xVbGv5GXZBh8rXNVTFvg7dzNRRW3GQIkZz9~nK72s9E6GAi8D5dhESRslGqOAK5xsC04EmFc4Mlk09WafNEVglXWSidDBfX5c4OU-4JEjqU9ZwFIV1h-DwQZ-mJ93eJfXhuTWaIA32pg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    },
-    mainImgUrl:
-      'https://s3-alpha-sig.figma.com/img/6a9e/e32d/49275dc65ca1d61672d87fe76a73c7c8?Expires=1696809600&Signature=fvAGyRZu-mR7MfmUXtgRt4cKk2Ku9X3uAFPI1oTVlGQ1VbVxn5qitVOUb2fD4WQuuNSsfZqH7SHJu0ZlEZ8BSjd40dr3a-j7mDq0pc9glR5NBxsZpqcgoKeUzOQyzwk2DxuGzWLe9tMhBJxonkfimRhggvQLXGTv7WnpUoaOKSkmVcvarr6UQvaXOdXeOtWVPqM98nzlLZURJ7g2qi-HACUM-cNtOmO5f6tZsyW1dL5WG09Q5Xg1-mOg8x5mHnCs3C4WxlOidxFPP3QoiMfEtiHIpWVPskN4n6HxIyeoJhgOBC-IwvV2CHDGVJqf8ub-Xzl~53Cu4CgD-stOdJZdCA__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    postId: 10,
-  },
-  {
-    id: 7,
-    content:
-      '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
-    createDate: String(new Date()),
-    spaceTitle: 'SPACE 401',
-    writer: {
-      nick: '최고양',
-      url: 'https://s3-alpha-sig.figma.com/img/02f9/c229/6e3726a13538fa71aab55c92c8eb950a?Expires=1696809600&Signature=F3aa5ljj2H4gtM9d~joIHMN1CZGJC8Nu5bomeA-a98TgrpJ20rsyPOOstLhG57YbuPX0GkuY5togwKHcSWUkYqi1w~aqrV-dBP4Sqhr2pC7JAUeSbwLLoGCQuzbIqEsKq4stgV3x24Xbk81asoGubOrYkzAB2CFWd5MoMMCbfyG3nSoy~dzhKQbQbAjUWmtaly-xlcdZe6xVbGv5GXZBh8rXNVTFvg7dzNRRW3GQIkZz9~nK72s9E6GAi8D5dhESRslGqOAK5xsC04EmFc4Mlk09WafNEVglXWSidDBfX5c4OU-4JEjqU9ZwFIV1h-DwQZ-mJ93eJfXhuTWaIA32pg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    },
-    mainImgUrl:
-      'https://s3-alpha-sig.figma.com/img/5931/fc27/51138b2ab3befac711602eb558ba0889?Expires=1696809600&Signature=H7TIVUPo0Py5XNCJpbiZCPITKuk3WDO-iAVIYqov44876FaQvjzH-W0pxeAXibmZulGzmzzZSxc0KZu5o8MinySRelZ0QEFQdwTSrMg4ZVvClTPVH8jes6dPDSkNEtB3sm5gElPrCsbCg~cb4H-R1Ovp1cgrOCImf8ErYnhumIXmLSM1EefJ6nAgCfkACrt4HGFpok~IZEQ5LSrBzrZqBqglctuSPtapIlSsdbYiGTiaFx-zeLTfl0ujHU-dLPceGayVu6bCitw89wLTX59umgZWEjQcW4Hezwe~~A5rYRqaFFDO2DkQ0ZWPYnndbPWM9RErH7JM5DiTtMxEei1QgA__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    postId: 10,
-  },
-  {
-    id: 8,
-    content:
-      '헐 여기 전에 가본데 자너 여기 좋음;; 헐 여기 전에 가본데 자너 여기 좋음;; ',
-    createDate: String(new Date()),
-    spaceTitle: 'SPACE 401',
-    writer: {
-      nick: '최고양',
-      url: 'https://s3-alpha-sig.figma.com/img/02f9/c229/6e3726a13538fa71aab55c92c8eb950a?Expires=1696809600&Signature=F3aa5ljj2H4gtM9d~joIHMN1CZGJC8Nu5bomeA-a98TgrpJ20rsyPOOstLhG57YbuPX0GkuY5togwKHcSWUkYqi1w~aqrV-dBP4Sqhr2pC7JAUeSbwLLoGCQuzbIqEsKq4stgV3x24Xbk81asoGubOrYkzAB2CFWd5MoMMCbfyG3nSoy~dzhKQbQbAjUWmtaly-xlcdZe6xVbGv5GXZBh8rXNVTFvg7dzNRRW3GQIkZz9~nK72s9E6GAi8D5dhESRslGqOAK5xsC04EmFc4Mlk09WafNEVglXWSidDBfX5c4OU-4JEjqU9ZwFIV1h-DwQZ-mJ93eJfXhuTWaIA32pg__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    },
-    mainImgUrl:
-      'https://s3-alpha-sig.figma.com/img/5337/9618/c6819ebc891905be191c8fc51d7fba08?Expires=1696809600&Signature=Kl9DjufuxFzmsLBrR4UnNSPkBRFuVGc~r0QeQ2YFoRBaTldUjnbIXFCDPD0pEWsvIWZ15qcUY-tlI5ZPp0vk-05vmh-JlHk0VfaLTDRSSX0bqqvBErzg0HRliJBmfyontnkwCbukqWelHms58RhnsDxE6HEuTW~IIq8w8lX9146Dq2lc8nL0trXvLHcQbSwsZtjm1rrgNlWAtSL7Rmj64397SsDVist9UXIcQNrd2r-RyJlPuBzuZDeBs3aoZBNETZgyrIkliHrZwOP7thldbCuoO0JAkgdrbBGtm4BfNmnfAzePsj42SqlnKBPWSKahbpJKsA1u1q7HXsEFuLagRw__&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4',
-    postId: 10,
-  },
 ];
 
 export default commentListMock;
