+++ client/src/router/Router.tsx
@@ -1,4 +1,4 @@
-import { createBrowserRouter } from 'react-router-dom';
+import {createBrowserRouter} from 'react-router-dom';
 import CreateSpace from '@/pages/CreateSpace';
 import CreatePost from '@/pages/CreatePost';
 import MainPage from '@/pages/Main';
@@ -9,70 +9,66 @@ import Landing from '@/pages/Landing';
 import SelectSpace from '@/pages/SelectSpace';
 import HeaderLayout from '@/layout/HeaderLayout/HeaderLayout';
 import Auth from '@/pages/Auth';
-import { PATH } from '@constants/path';
-import { Suspense } from 'react';
+import {PATH} from '@constants/path';
+import {Suspense} from 'react';
 
 const router = createBrowserRouter([
-  {
-    path: PATH.HOME,
-    element: <PrivateRoute />,
-    children: [
-      {
-        element: <HeaderLayout />,
+    {
+        path: PATH.HOME,
+        element: <PrivateRoute/>,
         children: [
-          /*스페이스*/
-          {
-            path: PATH.SPACE_SELECT,
-            element: <SelectSpace />,
-          },
-          {
-            path: PATH.SPACE + '/:spaceId',
-            element: (
-              <Suspense fallback={<></>}>
-                <MainPage />
-              </Suspense>
-            ),
-          },
-          {
-            path: PATH.SPACE_UPDATE + '/:spaceId',
-            // element: 업데이트 스페이스
-          },
-          {
-            path: PATH.POST_UPDATE,
-            // element: 업데이트 포스트
-          },
-          {
-            path: PATH.USER_INFO,
-            element: (
-              <Suspense fallback={<></>}>
-                <MyPage />
-              </Suspense>
-            ),
-          },
+            {
+                element: <HeaderLayout/>,
+                children: [
+                    /*스페이스*/
+                    {
+                        path: PATH.SPACE_SELECT,
+                        element: <SelectSpace/>,
+                    },
+                    {
+                        path: PATH.SPACE + '/:spaceId',
+                        element: (
+                            <Suspense fallback={<></>}>
+                                <MainPage/>
+                            </Suspense>
+                        ),
+                    },
+                    {
+                        path: PATH.SPACE_UPDATE + '/:spaceId',
+                        // element: 업데이트 스페이스
+                    },
+                    {
+                        path: PATH.POST_UPDATE,
+                        // element: 업데이트 포스트
+                    },
+                    {
+                        path: PATH.USER_INFO,
+                        element: <MyPage/>,
+                    },
+                ],
+            },
+            {
+                path: PATH.SPACE_CREATE,
+                element: <CreateSpace/>,
+            },
+            {
+                path: PATH.POST_CREATE,
+                element: <CreatePost/>,
+            },
         ],
-      },
-      {
-        path: PATH.SPACE_CREATE,
-        element: <CreateSpace />,
-      },
-      {
-        path: PATH.POST_CREATE,
-        element: <CreatePost />,
-      },
-    ],
-  },
-  {
-    path: PATH.LANDING,
-    element: <Landing />,
-  },
-  {
-    path: PATH.INVITE,
-    element: <InviteCode />,
-  },
-  {
-    path: PATH.AUTH,
-    element: <Auth />,
-  },
+    },
+    {
+        path: PATH.LANDING,
+        element: <Landing/>,
+    },
+    {
+        path: PATH.INVITE,
+        element: <InviteCode/>,
+    },
+    {
+        path: PATH.AUTH,
+        element: <Auth/>,
+    },
 ]);
 
 export default router;
